+++ lib/string-utils.ts
@@ -30,11 +30,22 @@ export function decamelize (str: string, joinString?: string): string {
   for (let i = 0; i < str.length; i++) {
     const chrLower = lowercase.charAt(i)
     const chrString = str.charAt(i)
-    if (chrLower !== chrString) {
+    if (chrLower !== chrString && i > 0) {
       notCamelcase += `${joinString}${lowercase.charAt(i)}`
     } else {
       notCamelcase += chrString
     }
   }
   return notCamelcase
+}
+
+export function looksLikeNumber (x: null | undefined | number | string): boolean {
+  if (x === null || x === undefined) return false
+  // if loaded from config, may already be a number.
+  if (typeof x === 'number') return true
+  // hexadecimal.
+  if (/^0x[0-9a-f]+$/i.test(x)) return true
+  // don't treat 0123 as a number; as it drops the leading '0'.
+  if (x.length > 1 && x[0] === '0') return false
+  return /^[-]?(?:\d+(?:\.\d*)?|\.\d+)(e[-+]?\d+)?$/.test(x)
 }