+++ rollup.config.js
@@ -1,4 +1,6 @@
+import cleanup from 'rollup-plugin-cleanup'
 import ts from '@wessberg/rollup-plugin-ts'
+import transformDefaultExport from 'ts-transform-default-export'
 
 const output = {
   format: 'cjs',
@@ -12,6 +14,14 @@ export default {
   input: './lib/index.ts',
   output,
   plugins: [
-    ts({ /* options */ })
+    ts({
+      transformers: ({ program }) => ({
+        afterDeclarations: transformDefaultExport(program)
+      })
+    }),
+    cleanup({
+      comments: 'none',
+      extensions: ['*']
+    })
   ]
 }